requirements.txt
autopep8==1.5.7
cycler==0.10.0
h5py==2.10.0
kiwisolver==1.3.1
matplotlib==3.4.2
numpy==1.20.3
opencv-contrib-python==4.5.2.54
pandas==1.2.5
Pillow==8.2.0
pycodestyle==2.7.0
pyparsing==2.4.7
pyrealsense2==2.47.0.3313
python-dateutil==2.8.1
pytz==2021.1
scipy==1.7.0
six==1.16.0
timm==0.4.9
toml==0.10.2
torch==1.7.1
torchvision==0.8.2
tqdm==4.56.0
typing-extensions==3.10.0.0
